requirements.txt
pdf2docx==0.5.6
python-docx==0.8.11
PyPDF2==3.0.1
pillow==10.0.0
flask==2.3.3
werkzeug==2.3.7
docx2pdf==0.1.8
gunicorn==21.2.0
python-dotenv==1.0.0
